Refactor MiqAeService into modules
To facilitate testing the modules in MiqAeService have been externalized
into separate files.
